requirements.txt
openai==1.13.0
google-auth==2.16.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.84.0
pandas==2.0.1
numpy==1.24.3
openpyxl==3.1.2
streamlit==1.32.0
pdf2image==1.16.3
Pillow==10.0.0
python-dotenv==1.0.0
reportlab==4.0.8